fixed DO2 displaying DO1

+++ main.py
@@ -152,7 +152,7 @@ class Controller(udi_interface.Node):
 
     def commandAO1(self, command):
 #        mynode = self.poly.getNode('controller')
-        mymessage = int(command.get("value"))
+        mymessage = float(command.get("value"))
         data = {}
         data['AO1'] = mymessage
         json_data = json.dumps(data)
